Update CSharp.md
Updated the first sample to make it easy to run. As for the second sample, it was for the Face - Identify API, which requires a preexisting group. So, I've replaced it with another sample that shows how to create a group. In the future, I can also add a third sample that then uses the Face - Identify API with the newly created group.

Note that in the Prerequisites section the topic points the reader to where they can obtain subscription keys. The samples include comments to direct the reader to replace the fake keys with their own keys.