Update CSharp.md
Updated the code to make it work after the reader simply copies and pastes it. The only thing they'll need to do is change the key. There's a comment to that effect in the code, above the line where the key is used. Also, in the prerequisites section there's a link to where they can obtain a key.