Update Java.md
Updated the samples to make them a bit easier to run after the user simply copies and pastes them. I replaced the Identify example with a Create-Group example because the latter lends itself better to a simple, standalone sample, whereas the former requires the creation of a person group first in order for it to work.